integrated postgres sql

+++ HMS/models.py
@@ -3,10 +3,10 @@ from django.db import models
 
 class Student(models.Model):
     name = models.CharField(max_length=200)
-    admission_no= models.IntegerField(primary_key=True)
+    admission_no= models.BigIntegerField(primary_key=True)
     hostel=models.CharField(max_length=50,default="null")
-    room_no=models.IntegerField(default=0)
-    phone_no=models.IntegerField(default=0000000000)
+    room_no=models.BigIntegerField(default=0)
+    phone_no=models.BigIntegerField(default=0000000000)
     address=models.CharField(max_length=100,default="null")
 
 class Login(models.Model):
@@ -17,14 +17,14 @@ class Complaints(models.Model):
     complaint_id=models.BigAutoField(primary_key=True)
     admission_no=models.ForeignKey(Student,on_delete=models.CASCADE)
     hostel=models.CharField(max_length=20,default='null')
-    room_no=models.IntegerField()
+    room_no=models.BigIntegerField()
     complaint_desc=models.TextField()
     status=models.BooleanField(default=False)
 
 class Workers(models.Model):
     name=models.CharField(max_length=25)
     designation=models.CharField(max_length=25)
-    phone_no=models.IntegerField(default=0)
+    phone_no=models.BigIntegerField(default=0)
